requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pyjwt==2.8.0
requests==2.31.0 